:recycle: Home, Profile 리팩토링 및 로그아웃 기능 버그 수정
