+++ src/routes/Auth.js
@@ -1,62 +1,31 @@
-import React, {useState} from 'react';
-import {authService, firebaseInstance} from "../fbase";
+import React from "react";
+import { authService, firebaseInstance } from "../fbase";
+import AuthForm from "../components/AuthForm";
 
 const Auth = () => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [newAccount, setNewAccount] = useState(true);
-  const [error, setError] = useState("");
-
-  const onChange = (event) => {
-    const {target: {name, value}} = event;
-    if (name === "email") {
-      setEmail(value)
-    } else if (name === "password") {
-      setPassword(value)
-    }
-  };
-
-  const onSubmit = async (event) => {
-    event.preventDefault();
-    try {
-      let data;
-      if(newAccount) {
-        data = await authService.createUserWithEmailAndPassword(email, password);
-      } else {
-        data = await authService.signInWithEmailAndPassword(email, password);
-      }
-      console.log(data)
-    } catch (error) {
-      setError(error.message)
-    }
-  };
-
-  const toggleAccount = () => setNewAccount(prev => !prev);
-
   const onSocialClick = async (event) => {
-    const {target: {name}} = event;
+    const {
+      target: { name },
+    } = event;
     let provider;
-    if(name === "google"){
-      provider = new firebaseInstance.auth.GoogleAuthProvider()
-    } else if(name === "github"){
-      provider = new firebaseInstance.auth.GithubAuthProvider()
+    if (name === "google") {
+      provider = new firebaseInstance.auth.GoogleAuthProvider();
+    } else if (name === "github") {
+      provider = new firebaseInstance.auth.GithubAuthProvider();
     }
-    const data = await authService.signInWithPopup(provider);
-    console.log(data);
+    await authService.signInWithPopup(provider);
   };
 
   return (
     <div>
-      <form onSubmit={onSubmit}>
-        <input name="email" type="text" placeholder="Email" required value={email} onChange={onChange}/>
-        <input name="password" type="password" placeholder="password" required value={password} onChange={onChange}/>
-        <input type="submit" value={newAccount ? "Create Account" : "Sign in"}/>
-        {error}
-      </form>
-      <span onClick={toggleAccount}>{newAccount ? "Sign in" : "Create Account"}</span>
+      <AuthForm />
       <div>
-        <button name="google" onClick={onSocialClick}>Continue with Google</button>
-        <button name="github" onClick={onSocialClick}>Continue with GitHub</button>
+        <button name="google" onClick={onSocialClick}>
+          Continue with Google
+        </button>
+        <button name="github" onClick={onSocialClick}>
+          Continue with GitHub
+        </button>
       </div>
     </div>
   );
